Fix minor regex issues in date dectection and move around files in models/crf package to avoid conflicts with django/app/models structure
